remove windows type
