+++ src/types.ts
@@ -1,29 +1,34 @@
 import {
   BoolType,
+  registerType,
+  typedef,
   DoubleType,
   FloatType,
   PaddingType,
-  registerType,
   StringType,
-  typedef,
 } from "./class-type";
 
 export const string_t = new StringType();
 export const padding_t = new PaddingType();
 
 // c-type
-export const char = registerType(1, false);
-export const bool = new BoolType(char);
-export const uchar = registerType(1);
-export const short = registerType(2, false);
+export const char = registerType<number, number>(1, false);
+export const uchar = registerType<number, number>(1);
 
-export const ushort = registerType(2);
-export const int = registerType(4, false);
-export const uint = registerType(4);
-export const long = registerType(4, false);
-export const ulong = registerType(4);
-export const longlong = registerType(8, false);
-export const ulonglong = registerType(8);
+export const short = registerType<number, number>(2, false);
+export const ushort = registerType<number, number>(2);
+
+export const int = registerType<number, number>(4, false);
+export const uint = registerType<number, number>(4);
+
+export const long = registerType<number, number>(4, false);
+export const ulong = registerType<number, number>(4);
+
+export const longlong = registerType<bigint, bigint>(8, false);
+export const ulonglong = registerType<bigint, bigint>(8);
+
+export const bool = new BoolType(char);
+
 export const float = new FloatType();
 export const double = new DoubleType();
 
@@ -37,28 +42,3 @@ export const uint8_t = typedef(uchar);
 export const uint16_t = typedef(ushort);
 export const uint32_t = typedef(uint);
 export const uint64_t = typedef(ulonglong);
-
-// windows
-export const BOOL = new BoolType(int);
-export const BYTE = typedef(uchar);
-export const WORD = typedef(ushort);
-export const DWORD = typedef(ulong);
-export const QWORD = registerType(8);
-
-export const FLOAT = typedef(float);
-export const DOUBLE = typedef(double);
-
-export const CHAR = typedef(char);
-export const UCHAR = typedef(uchar);
-
-export const SHORT = typedef(short);
-export const USHORT = typedef(ushort);
-
-export const INT = typedef(int);
-export const UINT = typedef(uint);
-
-export const LONG = typedef(long);
-export const ULONG = typedef(ulong);
-
-export const LONGLONG = typedef(longlong);
-export const ULONGLONG = typedef(ulonglong);
