+++ test/deeps.test.ts
@@ -1,8 +1,7 @@
-import { BOOL, int32_t, padding_t, string_t, StructBuffer } from "../src";
+import { int32_t, padding_t, string_t, StructBuffer, uint8_t } from "../src";
 import {
   bits,
   BitsType,
-  BoolType,
   PaddingType,
   StringType,
   StructType,
@@ -16,18 +15,13 @@ describe("deeps test", () => {
     expect((a as any).deeps).toEqual([]);
     expect((b as any).deeps).toEqual([2]);
     expect((c as any).deeps).toEqual([2, 3]);
-
     expect(a instanceof StructType).toEqual(true);
     expect(b instanceof StructType).toEqual(true);
     expect(c instanceof StructType).toEqual(true);
-
     expect(padding_t[2][4] instanceof PaddingType).toEqual(true);
-
     expect(bits(int32_t, {})[2][2] instanceof BitsType).toEqual(true);
-    expect(BOOL[2][2] instanceof BoolType).toEqual(true);
     expect(string_t[2][2] instanceof StringType).toEqual(true);
   });
-
   it("struct", () => {
     let a = new StructBuffer({});
     let b = a[2];
@@ -35,9 +29,18 @@ describe("deeps test", () => {
     expect((a as any).deeps).toEqual([]);
     expect((b as any).deeps).toEqual([2]);
     expect((c as any).deeps).toEqual([2, 3]);
-
     expect(a instanceof StructBuffer).toEqual(true);
     expect(b instanceof StructBuffer).toEqual(true);
     expect(c instanceof StructBuffer).toEqual(true);
+  });
+
+  it("deeps length", () => {
+    const type = uint8_t[2][3][4];
+    const obj = type.decode(new Array(type.length).fill(1));
+    // console.log(obj);
+
+    expect(obj.length).toBe(2);
+    expect(obj[0].length).toBe(3);
+    expect(obj[0][0].length).toBe(4);
   });
 });
